fix: tweaking file structures
