+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./button";
 export * from "./form";
 export * from "./modal";
+export * from "./notification-group";
